ai-secure-platform/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
httpx==0.27.0
python-docx==1.1.0
pdfplumber==0.10.4
drain3==0.9.11
scikit-learn==1.3.0
numpy==1.26.4
motor==3.3.2
PyJWT==2.8.0
python-dotenv==1.0.1
openai==1.16.2
passlib==1.7.4
bcrypt==4.0.1
PyMuPDF==1.24.1
python-jose[cryptography]==3.3.0
pytest==8.1.1
pytest-asyncio==0.23.6
slowapi==0.1.9
loguru==0.7.2
pydantic-settings==2.2.1
transformers>=4.38.0
sentence-transformers>=2.5.0
torch>=2.2.0
